+++ tests/functional.py
@@ -432,8 +432,9 @@ class WorkflowTest(unittest.TestCase):
             self.elsa_client.patch_collection(data={'status': 'to-sign'})
 
     def test_status_cannot_be_refresh_if_never_signed(self):
-        with self.assertRaises(KintoException):
+        with self.assertRaises(KintoException) as e:
             self.elsa_client.patch_collection(data={'status': 'to-refresh'})
+        assert "Collection never signed." in e.exception.message
 
     def test_review_can_be_cancelled_by_editor(self):
         create_records(self.client)